refactor: modify login & logout function with using useAuth.ts #41
